requirements.txt
pip~=22.3.1
attrs~=23.2.0
wheel~=0.38.4
cryptography~=41.0.7
Jinja2~=3.1.2
MarkupSafe~=2.1.3
werkzeug~=3.0.1
click~=8.1.7
blinker~=1.7.0
itsdangerous~=2.1.2
setuptools~=65.5.1
flask~=3.0.0
python-dotenv~=1.0.0
flask_dance